Almost finish notary.
There is still an error with the compilation of the contracts until it is possible to retrieve the
inserted devices. The problem appears after the compilation, when trying to retrieve the built contract
from 'build' directory. It seems that for DepositDevice, the compilation is skipping two methods
'getName' and 'getValue', which are not included in the contract definition.